Create README.md
Garuda Airline reservation System is a computerized system used to store and retrieve information and conduct transactions related to air travel. 
The project is aimed at exposing the relevance and importance of Airline Reservation Systems.
 It is projected towards enhancing the relationship between customers and airline agencies through the use of ARS's, and thereby making it convenient for the customers to book the flights as when they require such that they can utilize this software to make reservations.
 This software has two parts. First is user part and the administrator part. User part is used as a front end and administrator is the back end. 
Administrator is used by airline authority. It will allow the customers to access database and allow new customers to sign up for online access.